requirements.txt
discord.py
requests
json